move function definitions of template classes to header file
