+++ src/software/ai/hl/stp/play/play_base.h
@@ -17,7 +17,10 @@ public:
      *
      * @param ai_config_ptr shared pointer to ai_config
      */
-    explicit PlayBase(std::shared_ptr<TbotsProto::AiConfig> ai_config_ptr, bool requires_goalie);
+    explicit PlayBase(std::shared_ptr<TbotsProto::AiConfig> ai_config_ptr, bool requires_goalie)
+            : Play(ai_config_ptr, requires_goalie),
+              fsm{PlayFsm{ai_config_ptr}, PlaySubFsms{ai_config_ptr}...},
+              control_params(){}
 
     void updateTactics(const PlayUpdate &play_update) override =0;
